made the addingPopup layout like Figma. Added a checkbox to ItemUi

+++ app/src/main/java/com/cc221002/ccl3_tgf/data/model/SingleEntry.kt
@@ -18,7 +18,7 @@ data class SingleEntry(
 	val foodName: String,
 	val bbDate: String,
 	val categoryId: Int, // Reference to Category
-	val portionAmount: Int,
+	val portionAmount: Float,
 	val portionType: String,
 	val isChecked: Int,
 	@PrimaryKey(autoGenerate = true) val id: Int = 0,
